+++ internal/ontocli/root.go
@@ -36,6 +36,7 @@ func NewRootCmd() *cobra.Command {
 	root.AddCommand(initCmd())
 	root.AddCommand(newCmd())
 	root.AddCommand(advanceCmd())
+	root.AddCommand(bypassCmd())
 	root.AddCommand(closeCmd())
 	root.AddCommand(abandonCmd())
 	root.AddCommand(doctorCmd())